Delete the csv files (from Github), renamed and expanded upon the sentence ease script. Added a merge of the sentence ease with the tatoeba database with translations.
